deploy-tool/requirements.txt
boto3=1.26.136
pyyaml=6.0
